blockchainmanager

+++ core/src/main/java/net/consensys/liszt/core/common/RTransfer.java
@@ -13,8 +13,8 @@ public class RTransfer {
   public final PublicKey from;
   public final PublicKey to;
   public final BigInteger amount;
-  public final int rIdFrom;
-  public final int rIdTo;
+  public final short rIdFrom;
+  public final short rIdTo;
   public final Signature sig;
   public final Hash hash;
 
@@ -23,8 +23,8 @@ public class RTransfer {
       PublicKey from,
       PublicKey to,
       BigInteger amount,
-      int rIdFrom,
-      int rIdTo,
+      short rIdFrom,
+      short rIdTo,
       Signature sig) {
     this.nonce = nonce;
     this.from = from;
